Fix offset issues.
This commit is tested against all test files I have available. We tried
to parse the temperature array offset from the file to match that one
Autel device, but that ended up breaking other stuff. It seems the
offset is not explicitly recorded.

TODO: Figure out if they are using 273.0 or 273.15 K.